refactor(config): refactor config menu
