+++ src/views/Config/BasicConfig/Cluster.vue
@@ -1,26 +1,28 @@
 <template>
-  <div class="cluster">
-    <schema-form
-      path="/configs/cluster"
-      type="cluster"
-      :form="configs"
-      :btn-loading="saveLoading"
-      @save="handleSave"
-    >
-      <template #invite-node>
-        <el-form-item :label="$t('BasicConfig.invaiteNode')">
-          <p class="item-desc">{{ $t('BasicConfig.inviteNodeDesc') }}</p>
-          <invite-node></invite-node>
-        </el-form-item>
-      </template>
-    </schema-form>
+  <div class="cluster app-wrapper">
+    <el-card class="config-card">
+      <schema-form
+        path="/configs/cluster"
+        type="cluster"
+        :form="configs"
+        :btn-loading="saveLoading"
+        @save="handleSave"
+      >
+        <template #invite-node>
+          <el-form-item :label="$t('BasicConfig.invaiteNode')">
+            <p class="item-desc">{{ $t('BasicConfig.inviteNodeDesc') }}</p>
+            <invite-node></invite-node>
+          </el-form-item>
+        </template>
+      </schema-form>
+    </el-card>
   </div>
 </template>
 
 <script lang="ts">
 import { defineComponent, ref } from 'vue'
 import SchemaForm from '@/components/SchemaForm'
-import InviteNode from './InviteNode.vue'
+import InviteNode from './components/InviteNode.vue'
 import { getClusterConfigs, updateClusterConfigs } from '@/api/config'
 import { Cluster } from '@/types/config'
 import { ElMessage } from 'element-plus'
